cirrus: add POWER7/8/9 builds
These are compile-only since qemu's POWER support is buggy, but we
want to make sure that we get the right architecture versions when
adding POWER implementations so we don't call, for example, a function
which first appeared in POWER9 on POWER8.

Unfortunately this doesn't catch situations where we restrict a
function to a later ISA version when a prior version would work. Still,
a missed optimization is less of a problem than broken code.

This is disabled until I find a good solution for generating cross
files; I don't really like the idea of carrying more in the git repo
(I'd love to replace the ones in the Docker directory with a script,
too).